third_party/blink/web_tests/wpt_internal/html/canvas/drawElementImage/privacy/video-ignored.https.sub.html

<!DOCTYPE HTML>
<html>
<head>
  <title>drawElementImage does not draw cross-origin videos</title>
  <script src='/resources/testharness.js'></script>
  <script src='/resources/testharnessreport.js'></script>
  <script src='../support/getPixel.js'></script>
  <script src="/wpt_internal/resources/canvas-draw-element/waitForCanvasPaint.js"></script>
  <style>
  #child {
    width: 100px;
    height: 200px;
    background: red;
  }
  #child > video {
    width: 100px;
    height: 100px;
  }
  #sameOrigin {
    position: absolute;
    left: 0px;
    top: 0px;
  }
  #crossOrigin {
    position: absolute;
    left: 0px;
    top: 100px;
  }
  </style>
</head>
<body>
  <canvas id=canvas width="100" height="200" layoutsubtree>
    <div id=child>
      <video id=sameOrigin autoplay muted></video>
      <video id=crossOrigin autoplay muted></video>
    </div>
  </canvas>

  <script>
    onload = () => {
      promise_test(async function(t) {
        const sameOrigin = document.getElementById("sameOrigin");
        const crossOrigin = document.getElementById("crossOrigin");
        const child = document.getElementById("child");
        const canvas = document.getElementById("canvas");
        const videoPath = "wpt_internal/html/canvas/drawElementImage/resources/white.webm";
        const sameOriginUrl = `https://{{location[host]}}/${videoPath}`;
        const crossOriginUrl = `https://{{hosts[alt][www]}}:{{ports[https][0]}}/${videoPath}`;
        let loaded = false;

        await waitForCanvasPaint(canvas);
        return new Promise((resolve, reject) => {
          const context = canvas.getContext("2d");
          canvas.onpaint = () => {
            context.drawElementImage(child, 0, 0);
            if (loaded) {
              // Fetch all pixel data once to avoid multiple slow readbacks.
              const imgData = context.getImageData(0, 0, canvas.width, canvas.height);

              let pixel = getPixel(imgData, 50, 50);
              assert_array_equals(pixel, [255, 255, 255, 255], "Same origin should draw");

              pixel = getPixel(imgData, 50, 150);
              assert_array_equals(pixel, [255, 0, 0, 255], "Cross origin should not draw");
              resolve();
            }
          };

          const loadVideo = (el, url) => new Promise((resolve, reject) => {
            el.onloadeddata = resolve;
            el.onerror = () => reject(new Error(`Failed to load video : ${url}`));
            el.src = url;
          });
          Promise.all([
            loadVideo(sameOrigin, sameOriginUrl),
            loadVideo(crossOrigin, crossOriginUrl)
          ]).then(() => { loaded = true });
        });
      });
    };
  </script>
</body>
</html>
